changes for php frontend

+++ personal_hobby_program_frontend/public/php/src/app/hiking/hiking.component.ts
@@ -9,26 +9,31 @@ import { Hiking } from '../hikings/hikings.component';
   styleUrls: ['./hiking.component.css']
 })
 export class HikingComponent implements OnInit {
-  
-  alert:boolean=false;
-  hiking!:Hiking;
-  constructor(private router: ActivatedRoute ,private _hikingDataService:HikingsDataService, private _router:Router) {
-   }
 
-  hikingId = this.router.snapshot.params["routeId"];
-  ngOnInit(): void {  
-    this._hikingDataService.getHiking(this.hikingId).subscribe(hiking => {this.hiking=hiking});
+  alert: boolean = false;
+  hiking!: Hiking;
+  constructor(private router: ActivatedRoute, private _hikingDataService: HikingsDataService, private _router: Router) {
   }
-onEdit(hiking:Hiking):void{
-  this._router.navigate(["editRoutes/:hikingId"])
-}
 
-onDelete():void{
-  this._hikingDataService.deleteHikingData(this.hikingId).subscribe(()=>{
-    this.alert=true
-      setTimeout(() => {
-        this._router.navigate(["getRoutes"]);
-    }, 1000); 
-  })
-}
+  hikingId = this.router.snapshot.params["hikingId"];
+  ngOnInit(): void {
+    this._hikingDataService.getHiking(this.hikingId).subscribe(hiking => { this.hiking = hiking });
+  }
+  onEdit(hiking: Hiking): void {
+    this._router.navigate(["gethikings/:hikingId"])
+  }
+
+  onDelete(): void {
+    if (confirm("Are you sure to delete ")) {
+      this._hikingDataService.deleteHikingData(this.hikingId).subscribe(() => {
+        this.alert = true
+        setTimeout(() => {
+          this._router.navigate(["gethikings"])
+        }, 1000);
+        setTimeout(() => {
+          this.alert = false
+        }, 1000);
+      })
+    }
+  }
 }
